[JBTM-2203] Remove ReentrantLock from StateManager
