+++ ArjunaCore/txoj/classes/com/arjuna/ats/txoj/LockManager.java
@@ -33,7 +33,6 @@ package com.arjuna.ats.txoj;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.Enumeration;
 import java.util.concurrent.locks.ReentrantLock;
 
 import com.arjuna.ats.arjuna.ObjectModel;
@@ -712,12 +711,8 @@ public class LockManager extends StateManager
                 {
                     if (super.usingActions != null)
                     {
-                        Enumeration e = super.usingActions.elements();
-
-                        while (e.hasMoreElements())
+                        for (BasicAction action : super.usingActions.values())
                         {
-                            BasicAction action = (BasicAction) e.nextElement();
-
                             if (action != null)  // shouldn't be null!!
                             {
                                 /*
@@ -1320,6 +1315,66 @@ public class LockManager extends StateManager
             return unloadOk;
         }
     }
+
+    /**
+     * @return the mutex object used to lock this object.
+     * @since JTS 2.1.
+     */
+
+    protected final ReentrantLock getMutex ()
+    {
+        return mutex;
+    }
+
+    /**
+     * @return the result of the attempt to lock this object.
+     * @since JTS 2.1.
+     */
+
+    protected final boolean lockMutex ()
+    {
+        try
+        {
+            mutex.lock();
+            
+            return true;
+        }
+        catch (final Throwable ex)
+        {
+            return false;
+        }
+    }
+
+    /**
+     * @return the result of the attempt to unlock this object.
+     * @since JTS 2.1.
+     */
+
+    protected final boolean unlockMutex ()
+    {
+        try
+        {
+            mutex.unlock();
+            
+            return true;
+        }
+        catch (final Throwable ex)
+        {
+            return false;
+        }
+    }
+
+    /**
+     * @return <code>true</code> if the object was locked, <code>false</code> if
+     *         the attempt would cause the thread to block.
+     * @since JTS 2.1.
+     */
+
+    protected final boolean tryLockMutex ()
+    {
+        return mutex.tryLock();
+    }
+   
 
     protected String lockStoreType = txojPropertyManager.getTxojEnvironmentBean().getLockStoreType();
